backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis & Celery
redis==5.0.1
celery[redis]==5.3.6

# HTTP Client (for Yevveko API)
httpx==0.27.0

# File uploads
python-multipart==0.0.9
aiofiles==23.2.1

# Utils
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.27.0
